feat: recover download url format
